CHERI: Convert calls to free_page/free_pages
Calls to free_page and free_pages that cast its (pointer) argument
to unsigned long must cast it to a uintptr_t instead.

+++ arch/riscv/include/asm/pgalloc.h
@@ -132,7 +132,7 @@ static inline p4d_t *p4d_alloc_one(struct mm_struct *mm, unsigned long addr)
 static inline void __p4d_free(struct mm_struct *mm, p4d_t *p4d)
 {
 	BUG_ON((unsigned long)p4d & (PAGE_SIZE-1));
-	free_page((unsigned long)p4d);
+	free_page((uintptr_t)p4d);
 }
 
 #define p4d_free p4d_free
