+++ drivers/block/drbd/drbd_main.c
@@ -2467,8 +2467,8 @@ static int drbd_alloc_socket(struct drbd_socket *socket)
 
 static void drbd_free_socket(struct drbd_socket *socket)
 {
-	free_page((unsigned long) socket->sbuf);
-	free_page((unsigned long) socket->rbuf);
+	free_page((uintptr_t) socket->sbuf);
+	free_page((uintptr_t) socket->rbuf);
 }
 
 void conn_free_crypto(struct drbd_connection *connection)
