+++ drivers/block/ublk_drv.c
@@ -1975,7 +1975,7 @@ static void ublk_deinit_queue(struct ublk_device *ub, int q_id)
 	if (ubq->ubq_daemon)
 		put_task_struct(ubq->ubq_daemon);
 	if (ubq->io_cmd_buf)
-		free_pages((unsigned long)ubq->io_cmd_buf, get_order(size));
+		free_pages((uintptr_t)ubq->io_cmd_buf, get_order(size));
 }
 
 static int ublk_init_queue(struct ublk_device *ub, int q_id)
