+++ drivers/crypto/atmel-aes.c
@@ -2130,7 +2130,7 @@ static int atmel_aes_buff_init(struct atmel_aes_dev *dd)
 
 static void atmel_aes_buff_cleanup(struct atmel_aes_dev *dd)
 {
-	free_page((unsigned long)dd->buf);
+	free_page((uintptr_t)dd->buf);
 }
 
 static int atmel_aes_dma_init(struct atmel_aes_dev *dd)
