+++ drivers/infiniband/hw/cxgb4/mem.c
@@ -564,7 +564,7 @@ struct ib_mr *c4iw_reg_user_mr(struct ib_pd *pd, u64 start, u64 length,
 				mhp->wr_waitp);
 
 pbl_done:
-	free_page((unsigned long) pages);
+	free_page((uintptr_t) pages);
 	if (err)
 		goto err_pbl_free;
