+++ drivers/s390/block/dasd_eckd.c
@@ -5656,7 +5656,7 @@ static void dasd_eckd_dump_sense_ccw(struct dasd_device *device,
 		}
 		dasd_eckd_dump_ccw_range(device, from, last, page + len);
 	}
-	free_page((unsigned long) page);
+	free_page((uintptr_t) page);
 }
 
 
@@ -5771,7 +5771,7 @@ static void dasd_eckd_dump_sense_tcw(struct dasd_device *device,
 		sprintf(page + len, "SORRY - NO TSB DATA AVAILABLE\n");
 	}
 	dev_err(&device->cdev->dev, "%s", page);
-	free_page((unsigned long) page);
+	free_page((uintptr_t) page);
 }
 
 static void dasd_eckd_dump_sense(struct dasd_device *device,
@@ -6951,7 +6951,7 @@ dasd_eckd_init(void)
 		kfree(pe_handler_worker);
 		kfree(dasd_reserve_req);
 		kfree(dasd_vol_info_req);
-		free_page((unsigned long)rawpadpage);
+		free_page((uintptr_t)rawpadpage);
 	}
 	return ret;
 }
@@ -6962,7 +6962,7 @@ dasd_eckd_cleanup(void)
 	ccw_driver_unregister(&dasd_eckd_driver);
 	kfree(pe_handler_worker);
 	kfree(dasd_reserve_req);
-	free_page((unsigned long)rawpadpage);
+	free_page((uintptr_t)rawpadpage);
 }
 
 module_init(dasd_eckd_init);
