+++ drivers/s390/char/sclp_cmd.c
@@ -111,7 +111,7 @@ int _sclp_get_core_info(struct sclp_core_info *info)
 	}
 	sclp_fill_core_info(info, sccb);
 out:
-	free_pages((unsigned long) sccb, get_order(length));
+	free_pages((uintptr_t) sccb, get_order(length));
 	return rc;
 }
 
@@ -216,7 +216,7 @@ static int do_assign_storage(sclp_cmdw_t cmd, u16 rn)
 		break;
 	}
 out:
-	free_page((unsigned long) sccb);
+	free_page((uintptr_t) sccb);
 	return rc;
 }
 
@@ -274,7 +274,7 @@ static int sclp_attach_storage(u8 id)
 		break;
 	}
 out:
-	free_page((unsigned long) sccb);
+	free_page((uintptr_t) sccb);
 	return rc;
 }
 
@@ -528,7 +528,7 @@ static int __init sclp_detect_standby_memory(void)
 		goto out;
 	sclp_add_standby_memory();
 out:
-	free_page((unsigned long) sccb);
+	free_page((uintptr_t) sccb);
 	return rc;
 }
 __initcall(sclp_detect_standby_memory);
@@ -578,7 +578,7 @@ static int do_chp_configure(sclp_cmdw_t cmd)
 		break;
 	}
 out:
-	free_page((unsigned long) sccb);
+	free_page((uintptr_t) sccb);
 	return rc;
 }
 
@@ -649,6 +649,6 @@ int sclp_chp_read_info(struct sclp_chp_info *info)
 	memcpy(info->standby, sccb->standby, SCLP_CHP_INFO_MASK_SIZE);
 	memcpy(info->configured, sccb->configured, SCLP_CHP_INFO_MASK_SIZE);
 out:
-	free_page((unsigned long) sccb);
+	free_page((uintptr_t) sccb);
 	return rc;
 }
