+++ drivers/thunderbolt/debugfs.c
@@ -74,7 +74,7 @@ static void *validate_and_copy_from_user(const void __user *user_buf,
 
 	nbytes = min_t(size_t, *count, PAGE_SIZE);
 	if (copy_from_user(buf, user_buf, nbytes)) {
-		free_page((unsigned long)buf);
+		free_page((uintptr_t)buf);
 		return ERR_PTR(-EFAULT);
 	}
 
@@ -162,7 +162,7 @@ static ssize_t regs_write(struct tb_switch *sw, struct tb_port *port,
 out:
 	pm_runtime_mark_last_busy(&sw->dev);
 	pm_runtime_put_autosuspend(&sw->dev);
-	free_page((unsigned long)buf);
+	free_page((uintptr_t)buf);
 
 	return ret < 0 ? ret : count;
 }
@@ -299,7 +299,7 @@ margining_ber_level_write(struct file *file, const char __user *user_buf,
 	usb4->margining->ber_level = val;
 
 out_free:
-	free_page((unsigned long)buf);
+	free_page((uintptr_t)buf);
 out_unlock:
 	mutex_unlock(&tb->lock);
 
@@ -442,7 +442,7 @@ margining_lanes_write(struct file *file, const char __user *user_buf,
 	mutex_unlock(&tb->lock);
 
 out_free:
-	free_page((unsigned long)buf);
+	free_page((uintptr_t)buf);
 	return ret < 0 ? ret : count;
 }
 
@@ -515,7 +515,7 @@ static ssize_t margining_mode_write(struct file *file,
 	mutex_unlock(&tb->lock);
 
 out_free:
-	free_page((unsigned long)buf);
+	free_page((uintptr_t)buf);
 	return ret ? ret : count;
 }
 
@@ -772,7 +772,7 @@ static ssize_t margining_test_write(struct file *file,
 	mutex_unlock(&tb->lock);
 
 out_free:
-	free_page((unsigned long)buf);
+	free_page((uintptr_t)buf);
 	return ret ? ret : count;
 }
 
@@ -840,7 +840,7 @@ static ssize_t margining_margin_write(struct file *file,
 	mutex_unlock(&tb->lock);
 
 out_free:
-	free_page((unsigned long)buf);
+	free_page((uintptr_t)buf);
 	return ret ? ret : count;
 }
 
@@ -1086,7 +1086,7 @@ static ssize_t counters_write(struct file *file, const char __user *user_buf,
 out:
 	pm_runtime_mark_last_busy(&sw->dev);
 	pm_runtime_put_autosuspend(&sw->dev);
-	free_page((unsigned long)buf);
+	free_page((uintptr_t)buf);
 
 	return ret < 0 ? ret : count;
 }
