+++ drivers/vhost/vdpa.c
@@ -1203,7 +1203,7 @@ out:
 unlock:
 	mmap_read_unlock(dev->mm);
 free:
-	free_page((unsigned long)page_list);
+	free_page((uintptr_t)page_list);
 	return ret;
 
 }
